fix inspector and refactor to use_query_params
