+++ assets/src/components/v2/screen_page.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useParams } from "react-router-dom";
 import ScreenContainer from "Components/v2/screen_container";
 import { ScreenIDProvider } from "Hooks/v2/use_screen_id";
-import { getQueryParamMap } from "Util/query_params";
+import { useQueryParams } from "Hooks/v2/use_query_params";
 
 interface ScreenPageProps {
   id?: string;
@@ -12,7 +12,7 @@ interface ScreenPageProps {
 const ScreenPage = ({ id, paramKeys }: ScreenPageProps) => {
   const screenId = id ?? (useParams() as { id: string }).id;
 
-  const queryParams = paramKeys ? getQueryParamMap(paramKeys) : undefined;
+  const queryParams = useQueryParams(paramKeys ?? []);
 
   return (
     <ScreenIDProvider id={screenId}>
